move black hole postprocess to black hole folder
